Fixed compilation issues

+++ includes/map.h
@@ -6,7 +6,7 @@
 /*   By: alexphil <alexphil@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/01 15:30:29 by alexphil          #+#    #+#             */
-/*   Updated: 2024/02/06 16:31:29 by alexphil         ###   ########.fr       */
+/*   Updated: 2024/02/06 16:40:14 by alexphil         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,7 +66,7 @@ int		process_ambient(t_import *import, char **infos);
 
 // Process Utils
 int		set_light(float light, char *info);
-int		set_rgb(int *rgb, char *info);
+int		set_rgb(float *rgb, char *info);
 int		set_xyz(float *xyz, char *info);
 int		set_normal(float *normal, char *info);
 int		set_fov(int	fov, char *info);
